Adding documentation for the ALPS Consumer.
