+++ profiles/alps/constructor.go
@@ -20,22 +20,9 @@ type Constructor interface {
 	AddLink(ConstructorLink)
 }
 
-type Doc interface {
-	// Href returns the Href of the documentation
-	Href() url.URL
-	Format() string
-	Value() string
-
-	json.Unmarshaler
-}
-
 type ConstructorDoc interface {
 	Doc
 	json.Marshaler
-}
-
-type Descriptor interface {
-	json.Unmarshaler
 }
 
 type ConstructorDescriptor interface {
@@ -43,23 +30,9 @@ type ConstructorDescriptor interface {
 	json.Marshaler
 }
 
-type Ext interface {
-	Id() string
-	Href() url.URL
-	Value() string
-	json.Unmarshaler
-}
-
 type ConstructorExt interface {
 	Ext
 	json.Marshaler
-}
-
-type Link interface {
-	Href() url.URL
-	Rel() string
-
-	json.Unmarshaler
 }
 
 type ConstructorLink interface {
